backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
Pillow
jax
jaxlib
flax
msgpack